WIP: Update launch tests
