+++ gisnav/gisnav/__init__.py
@@ -13,7 +13,6 @@ ROS namespace and :term:`core` node names are hard-coded inside the static
 public node entrypoints defined here. Other node initialization arguments are p
 rovided via ROS 2 launch arguments.
 """
-import cProfile
 import io
 import pstats
 from typing import Optional
@@ -41,11 +40,12 @@ def _run(constructor: rclpy.node.Node, *args, **kwargs):
     :param **kwargs: Node constructor kwargs
     :return:
     """
-    if __debug__:
-        profile = cProfile.Profile()
-        profile.enable()
-    else:
-        profile = None
+    # if __debug__:
+    #    profile = cProfile.Profile()
+    #    profile.enable()
+    # else:
+    #    profile = None
+    profile = None
 
     node: Optional[Node] = None
     try:
